+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - `Profile.ThickenedEdgePolygons`
 - `Elements.MEP`
 - `GeometricElement.RepresentationInstances`
+- Custom implementations of `ToPolyline` for `IndexedPolycurve`, `Polyline` and `Polygon`.
 
 ### Fixed
 
@@ -58,6 +59,7 @@
 - `Polyline.Frames` now works correctly with `startSetbackDistance` and `endSetbackDistance` parameters.
 - `Polygon.Frames` now works correctly with `startSetbackDistance` and `endSetbackDistance` parameters.
 - `BoundedCurve.ToPolyline` now works correctly for `EllipticalArc` class.
+- `Vector3.DistanceToDistanceTo(Polygon polygon, out Vector3 closestPoint)` now also checks the last edge of the polygon.
 
 
 ### Changed